refactor(txn hash): move to separate mod & add padding test
